2_openai/requirements.txt
# Requirements for Research Assistant - Hugging Face Spaces

# Core dependencies
gradio>=5.0.0
python-dotenv>=1.0.0

# OpenAI Agents SDK
openai-agents>=0.0.15
openai>=1.68.0

# For deployment
pydantic>=2.0.0

